+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+* [THORN-207] Add HW readback for channeliser truncation.
+* [SKB-999] Correct PreAduPresent property to account for both preadu.renamed to array property PreAduFitted (default [True, True])
+* [SKB-955] Correct transient transition to ON when MccsTile is turning OFF.
+
+## 7.3.0
+
+* [THORN-218] Added 51 monitoring points as tango device attributes in tile. (ADC temperatures, FE currents and voltages)
 * [THORN-241] Add cleanup code to delete_device. This fixes an issue where multiple calls to init() lead to multiple polling threads.
 * [SKB-928] Remove configuration attributes from polling. Configuration is now 'read_on_connect' or 'read_on_change'
 * [SKB-928] SpsStation staticTimeDelay now raised RuntimeError when information for TPM mapping not present